split balance and transfer function + run transfer tests only on localhost
